refactor(quiz): ordenando componentes para blogs en carpeta blog
