+++ app/posts/blog/create/[communityId]/page.tsx
@@ -5,9 +5,9 @@ import { useParams, useRouter } from "next/navigation";
 import { Input } from "@headlessui/react";
 import { useForm, Controller } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import BlogPreview from "../../components/BlogPreview";
-import BlogHeader from "../../components/CreateBlogHeader";
-import Tiptap from "../../components/TipTap";
+import BlogPreview from "../../components/blog/BlogPreview";
+import BlogHeader from "../../components/blog/CreateBlogHeader";
+import Tiptap from "../../components/blog/TipTap";
 import { BlogDraft, BlogPost } from "@/types/posts/blog";
 import Modal from "@/components/shared/BlogModal";
 import { createBlogPost } from "@/services/posts/blogs/blogsService";
